+++ src/personal_info/info/personal_info_5_b2w2.rs
@@ -132,7 +132,7 @@ impl PersonalFormInfo for PersonalInfo5B2W2<'_> {
 }
 
 impl PersonalAbility for PersonalInfo5B2W2<'_> {
-    fn get_index_of_ability(&self, ability_id: u16) -> Option<usize> {
+    fn get_index_of_ability(&self, ability_id: u16) -> Option<u8> {
         if ability_id == self.ability_1() {
             Some(0)
         } else if ability_id == self.ability_2() {
