+++ src/personal_info/info/personal_info_9_sv.rs
@@ -144,7 +144,7 @@ impl PersonalFormInfo for PersonalInfo9SV<'_> {
 }
 
 impl PersonalAbility for PersonalInfo9SV<'_> {
-    fn get_index_of_ability(&self, ability_id: u16) -> Option<usize> {
+    fn get_index_of_ability(&self, ability_id: u16) -> Option<u8> {
         if ability_id == self.ability_1() {
             Some(0)
         } else if ability_id == self.ability_2() {
