+++ src/pkm/traits/metadata/game_value_limit.rs
@@ -3,7 +3,7 @@ pub trait GameValueLimit {
     fn max_move_id(&self) -> u16;
     fn max_item_id(&self) -> u16;
     fn max_ability_id(&self) -> u16;
-    fn max_ball_id(&self) -> u16;
+    fn max_ball_id(&self) -> u8;
     fn max_game_id(&self) -> u8;
     fn min_game_id(&self) -> u8;
     fn max_iv(&self) -> u8;
